feat: Prim attribute values table is now functional.
This commit adds functionality to the prim attribute value table,
displaying the value of the currently selected attribute in the
prim attributes table.

In this commit I've made sure to explicitly highlight that what's
displayed are prim attributes and not properties, support for properties
(attributes and relationships) will come in a later iteration. This
commit also removes all Placeholder widgets.

Finally, I've replaced the usage of automatic event handler methods
to explicitly defining them using the on decorator as I was running
into widgets that subclass the same parent widget, which caused them
to send the same messages and trigger the same event handler.