feat: PrimCompositionList is now interactive.
This commit introduces a change that I'm really happy about, selecting
a layer for a prim in the PrimCompositionList widget will make the
attributes data table show the attributes for the prim spec in that
layer, essentially allowing you to inspect the state of the prim
in that layer directly.

There are many things to improve in the current iteration and I'm
sure that the implementation could be better, but at least it serves
as a proof of concept of what's possible.