fix: PrimCompositionList is now PrimLayerStackTable
I introduced this change as using a table seemed more appropriate
for representing a stack on the UI, having rows allows me to add
more data to them in the future, using a table also gets rid of a
visual bug I had when clearing the table, when I used a ListView
the table would flicker. The new naming is more appropriate as the
widget is showing a stack of layers.

I added a column called "Specifier" which displays the name of the
specifier for the prim spec in that layer. I plan on adding
additional data to be displayed in different columns.