feat: Prim Properties are now color tagged
This commit introduces color tagging for properties listed in in the
PrimPropertiesTable:

- Each layer in the PrimCompositionTable is assigned a color based on
its hash.
- Each property name in the PrimPropertiesTable is colored based on the
hash color of the first layer in its layer stack (Winning opinion).
- If we are seeing a PropertySpec, they take the color of the layer that
owns it.
- The type of a property is now colored to represent a type of value:
    - Gray: No Value.
    - Amber: Fallback Value.
    - Cyan: Default Value.
    - Green: Time Sampled Value.

Fixes introduced:
- When a PropertySpec was selected in the table the Values table was
displaying the value of the final composed property and not the value of
the property spec, the same was happening for the Property Metadata
table, now they display the correct values.