fix: Concretized dependency versions. Set USD version to 24.5
I decided to concretize USD to that particular version as it's the
latest for which there have been stubs released. I wanted to use a
change from 24.11 that made TimeSamples data accessible directly
through an AttributeSpec instead of  through its layer but I was
surprised to see that the stubs did not recognize the new methods
but they worked at runtime, that's when I discovered that there
were no stubs for 24.11 yet.